Move asset master-data model to dedicated Assets feature #474
The Asset, AssetListing, MarketDataSymbol, AssetIdentifier and PriceQuote
entities (with their enums) and master-data repositories are not user
transactions, so they no longer live under FinancialAccounts. They move to a
new top-level FinanceManager.Domain.Assets feature.

InvestmentTransaction (and its repository) stay under
FinancialAccounts.Investments and link to an AssetListing. The dependency is
now one-way: the Assets feature never references user-transaction types
(dropped AssetListing.Transactions; the FK is defined from the transaction
side). Regenerated the AddInvestmentModel migration/snapshot to match the new
namespaces; the database schema is unchanged.
