+++ code/FinanceManager.Domain/Assets/Entities/AssetIdentifier.cs
@@ -1,4 +1,4 @@
-namespace FinanceManager.Domain.FinancialAccounts.Investments.Entities;
+namespace FinanceManager.Domain.Assets.Entities;
 
 /// <summary>
 /// Represents an additional identifier assigned to an <see cref="Asset"/>. Not every instrument
