+++ code/FinanceManager.Domain/Assets/Entities/AssetIdentifierType.cs
@@ -1,4 +1,4 @@
-namespace FinanceManager.Domain.FinancialAccounts.Investments.Entities;
+namespace FinanceManager.Domain.Assets.Entities;
 
 /// <summary>
 /// The kind of identifier stored in an <see cref="AssetIdentifier"/>.
