+++ code/FinanceManager.Domain/Assets/Entities/AssetType.cs
@@ -1,4 +1,4 @@
-namespace FinanceManager.Domain.FinancialAccounts.Investments.Entities;
+namespace FinanceManager.Domain.Assets.Entities;
 
 /// <summary>
 /// Classifies the kind of financial instrument an <see cref="Asset"/> represents.
