+++ code/FinanceManager.Domain/Assets/Entities/DistributionPolicy.cs
@@ -1,4 +1,4 @@
-namespace FinanceManager.Domain.FinancialAccounts.Investments.Entities;
+namespace FinanceManager.Domain.Assets.Entities;
 
 /// <summary>
 /// ETF/fund dividend handling: whether income is reinvested (accumulating) or paid out (distributing).
