+++ code/FinanceManager.Domain/Assets/Entities/IdentifierScope.cs
@@ -1,4 +1,4 @@
-namespace FinanceManager.Domain.FinancialAccounts.Investments.Entities;
+namespace FinanceManager.Domain.Assets.Entities;
 
 /// <summary>
 /// Describes which level an <see cref="AssetIdentifier"/> applies to, because the same identifier
