+++ code/FinanceManager.Domain/Assets/Entities/PriceQuote.cs
@@ -1,4 +1,4 @@
-namespace FinanceManager.Domain.FinancialAccounts.Investments.Entities;
+namespace FinanceManager.Domain.Assets.Entities;
 
 /// <summary>
 /// Represents a cached price returned by a market data provider for an <see cref="AssetListing"/>.
