+++ code/FinanceManager.Domain/Assets/Entities/ReplicationMethod.cs
@@ -1,4 +1,4 @@
-namespace FinanceManager.Domain.FinancialAccounts.Investments.Entities;
+namespace FinanceManager.Domain.Assets.Entities;
 
 /// <summary>
 /// ETF replication method describing how the fund tracks its benchmark index.
