+++ code/FinanceManager.Infrastructure/Contexts/Configurations/InvestmentTransactionConfiguration.cs
@@ -1,3 +1,4 @@
+using FinanceManager.Domain.Assets.Entities;
 using FinanceManager.Domain.FinancialAccounts.Investments.Entities;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
@@ -18,6 +19,13 @@ public class InvestmentTransactionConfiguration : IEntityTypeConfiguration<Inves
         builder.Property(x => x.Notes).HasMaxLength(512);
 
         builder.Ignore(x => x.SignedQuantity);
+
+        // Transactions link to an asset listing; the Assets feature does not back-reference user transactions.
+        // Keep transactions if a listing is ever removed by requiring the listing to be transaction-free first.
+        builder.HasOne(x => x.AssetListing)
+            .WithMany()
+            .HasForeignKey(x => x.AssetListingId)
+            .OnDelete(DeleteBehavior.Restrict);
 
         builder.HasIndex(x => x.UserId);
         builder.HasIndex(x => x.AssetListingId);
