+++ code/FinanceManager.Infrastructure/Contexts/Configurations/MarketDataSymbolConfiguration.cs
@@ -1,4 +1,4 @@
-using FinanceManager.Domain.FinancialAccounts.Investments.Entities;
+using FinanceManager.Domain.Assets.Entities;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
